+++ src/__init__.py
@@ -2,7 +2,7 @@
 from .client import Client
 
 
-VERSION = (0, 0, 7)
+VERSION = (0, 0, 8)
 __version__ = '.'.join(str(x) for x in VERSION)
 
 __all__ = ['Client']